requirements.txt
Django==5.2.3
graphene-django==3.2.2
django-graphql-jwt==0.4.0
psycopg[binary]
python-decouple==3.8
django-cors-headers==4.3.1
cloudinary==1.40.0
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
requests==2.32.3
Pillow==10.4.0
unidecode==1.4.0
channels==4.0.0
channels-redis==4.1.0
uvicorn[standard]==0.32.1
daphne==4.0.0
bleach==6.0.0
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
websockets==12.0
